\name{check.variant}
\alias{check.variant}
\title{Check zvariant object's attributes for consistency.}
\usage{
  check.variant(object)
}
\arguments{
  \item{object}{of class zvariant}
}
\value{
  A boolean value TRUE if everything is ok, otherwise a
  character vector of encounterd errors. All encountered
  warning are printed on-the-fly.
}
\description{
  Check zvariant object's attributes for consistency.
}
\note{
  For package internal use only
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\keyword{internal}

