% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{check_variant}
\alias{check_variant}
\title{Check zvariant object's attributes for consistency.}
\usage{
check_variant(object)
}
\arguments{
\item{object}{of class zvariant}
}
\value{
A boolean value TRUE if everything is ok, otherwise a character
  vector of encounterd errors. All encountered warnings are printed
  on-the-fly.
}
\description{
Check zvariant object's attributes for consistency.
}
\note{
For package internal use only
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\keyword{internal}
