\name{clean.str}
\alias{clean.str}
\title{Clean leading and trailing whitespaces from a given string. Additionally,
all occurrences of multiple whitespaces are replaced with a single
whitespace.}
\usage{
  clean.str(x)
}
\arguments{
  \item{x}{Character string.}
}
\value{
  An absolute path to a file of NULL if the path does not
  exist
}
\description{
  Clean leading and trailing whitespaces from a given
  string. Additionally, all occurrences of multiple
  whitespaces are replaced with a single whitespace.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

