% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{clean_str}
\alias{clean_str}
\title{Clean leading and trailing whitespaces from a given string. Additionally,
all occurrences of multiple whitespaces are replaced with a single
whitespace.}
\usage{
clean_str(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
An absolute path to a file of NULL if the path does not exist.
}
\description{
Clean leading and trailing whitespaces from a given string. Additionally,
all occurrences of multiple whitespaces are replaced with a single
whitespace.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

