% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decimalplaces}
\alias{decimalplaces}
\title{Find out the number of decimal places in a number.}
\usage{
decimalplaces(x, true_number = FALSE)
}
\arguments{
\item{x}{Float or double numeric number.}

\item{true_number}{Logical setting whether the true number (see notes) of
decimal places.}
}
\value{
Integer number of decimal places. Maximum
}
\description{
Original implementation from https://stackoverflow.com/questions/5173692/how-to-return-number-of-decimal-places-in-r
}
\note{
R usually restricts the number of decimal to 9 in printing etc. Unless
\code{true_number = TRUE}, return 9 and give a warning.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
