% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{file_path_relative_to}
\alias{file_path_relative_to}
\title{Transform an absolute path to relative path in relation to given location}
\usage{
file_path_relative_to(path, org_relative_to, new_relative_to)
}
\arguments{
\item{path}{Character string path.}

\item{org_relative_to}{Character string path to which \code{path} is originally
relative to.}

\item{new_relative_to}{Character string path to which \code{path} is supposed to
be relative to.}
}
\value{
Character string relative file path.
}
\description{
Transform an absolute path to relative path in relation to given location
}
\note{
Both \code{path} and \code{relative_to} must be in absolute form.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
