\name{get.tutorialdir}
\alias{get.tutorialdir}
\title{Get the directory of Zonation tutorial.}
\usage{
  get.tutorialdir()
}
\arguments{
  \item{\code{x}}{character string directory path.}
}
\value{
  \code{path} Character path to Zonation tutorial
  directory.
}
\description{
  Get the directory of Zonation tutorial.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

