\docType{methods}
\name{getVariant}
\alias{getVariant}
\title{getVariant
Get a specified variant in a Zonation project}
\arguments{
  \item{x}{Zproject object}

  \item{index}{int or string index defining the variant
  required}
}
\value{
  Zvariant object
}
\description{
  getVariant Get a specified variant in a Zonation project
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

