\name{initialize}
\alias{initialize}
\title{Create an instance of the Zproject class using new/initialize.}
\arguments{
  \item{root}{Character string path to the root of the
  project (must exist).}

  \item{name}{Character string naming the variant.}

  \item{bat.file}{Zonation specific batch (.bat) file to
  read the variant specifics from (must exist).}
}
\description{
  Create an instance of the Zproject class using
  new/initialize.

  Create an instance of the Zvariant class using
  new/initialize.
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>

  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
  \code{\link{initialize}}

  \code{\link{initialize}}
}
\keyword{internal}

