% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{load_zproject}
\alias{load_zproject}
\title{Load a Zonation project.}
\usage{
load_zproject(root, debug = FALSE)
}
\arguments{
\item{root}{Character string path pointing to an existing directory
(with potentially bat-files in it) or to a new directory to be created.}

\item{debug}{logical defining if debugging level for logging should be used.}
}
\value{
A \code{Zproject} object.
}
\description{
Loads an existing Zonation project as an object of
\code{\link[zonator:Zproject-class]{Zproject-class}}. Individual variants
within the Zonation project are parsed into
\code{\link[zonator:Zvariant-class]{Zvariant-class}} objects and potential
results into \code{\link[zonator:Zresults-class]{Zresults-class}} objects.
}
\seealso{
\code{\link[zonator:Zproject-class]{Zproject-class}},
  \code{\link[zonator:Zvariant-class]{Zvariant-class}} and
  \code{\link[zonator:create_zproject]{create_zproject}}
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
