% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{names,Zproject-method}
\alias{names,Zproject-method}
\title{Names of variants in Zproject}
\usage{
\S4method{names}{Zproject}(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
Get the names of all the variants within a given \code{\link{Zproject}}.
}

