\name{nvariants}
\alias{nvariants}
\title{nvariants
Get the number of variants included in a Zonation project}
\arguments{
  \item{x}{Zproject object}
}
\value{
  int number of variants
}
\description{
  nvariants Get the number of variants included in a
  Zonation project
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

