\name{open.dir}
\alias{open.dir}
\title{open.dir
Open the directory of a Zproject using the system file browser.}
\arguments{
  \item{x}{object}
}
\value{
  invisible
}
\description{
  Currently support Windows Explorer (Windows) amd Dolphin
  (Linux/KDE)
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

