\docType{methods}
\name{opendir}
\alias{opendir}
\alias{opendir,Zproject-method}
\title{Open the directory of a Zproject using the system file browser.}
\usage{
opendir(object)

\S4method{opendir}{Zproject}(object)
}
\arguments{
  \item{x}{object.}
}
\value{
invisible
}
\description{
Currently support Windows Explorer (Windows) amd Dolphin
(Linux/KDE).
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
\code{\link{Zproject-class}} and
\code{\link{Zvariant-class}}
}

