% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zresults-methods.R,
%   R/zvariant-methods.R
\name{outdir}
\alias{outdir}
\alias{outdir,Zresults-method}
\alias{outdir,Zvariant-method}
\title{Get path to output directory.}
\usage{
outdir(x)

\S4method{outdir}{Zresults}(x)

\S4method{outdir}{Zvariant}(x)
}
\arguments{
\item{x}{\code{\link[zonator]{Zvariant-class}} or
\code{\link[zonator:Zresults-class]{Zresults}} object.}
}
\value{
character string path to output directory location.
}
\description{
Zonation variant has an output directory defined in project bat-file. This
is of course the same path as in the results of the particular variant.
}
\seealso{
\code{\link{Zvariant-class}} \code{\link{Zresults-class}}
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
