\name{read.ini}
\alias{read.ini}
\title{read.ini
Read a Windows-style ini-file that is for configuration information.}
\usage{
  read.ini(infile)
}
\arguments{
  \item{infile}{character string input file path}
}
\value{
  a list of parsed ini-parameters
}
\description{
  read.ini Read a Windows-style ini-file that is for
  configuration information.
}
\note{
  Adapted from http://bit.ly/11e4Jh0
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

