\name{read.ppa.lsm}
\alias{read.ppa.lsm}
\title{read.ppa.lsm}
\usage{
  read.ppa.lsm(x)
}
\arguments{
  \item{x}{input file path}
}
\value{
  list of 3 data frames.
}
\description{
  Read Zonation post-processing analysis (ppa) result file
  and return a list of 3 items: 1. Most important species
  in units x 2. Average proportion remaining over all spp
  in units 3. Data fractions in units
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

