\name{read.spp}
\alias{read.spp}
\title{Read Zonation variant specific spp-file.}
\usage{
  read.spp(infile)
}
\arguments{
  \item{infile}{Character string input file path.}
}
\value{
  data.frame of parsed spp data.
}
\description{
  Read Zonation variant specific spp-file.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

