% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-input.R
\name{read_dat}
\alias{read_dat}
\title{Read a dat file (Windows-style ini-file) for configuration information.}
\usage{
read_dat(infile)
}
\arguments{
\item{infile}{Character string input file path.}
}
\value{
List of parsed ini-parameters.
}
\description{
Read a dat file (Windows-style ini-file) for configuration information.
}
\note{
Adapted from http://bit.ly/11e4Jh0
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
