% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-input.R
\name{read_features_info}
\alias{read_features_info}
\title{Read a features info file.}
\usage{
read_features_info(infile)
}
\arguments{
\item{infile}{Character string input file path.}
}
\value{
Data frame of parsed features info data.
}
\description{
Read a features info file.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
