% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-input.R
\name{read_groups}
\alias{read_groups}
\title{Read a groups file.}
\usage{
read_groups(infile)
}
\arguments{
\item{infile}{Character string input file path.}
}
\value{
Data frame of parsed groups parameters.
}
\description{
Read a groups file.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
