% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-input.R
\name{read_ini}
\alias{read_ini}
\title{Read a Windows-style ini-file that is for configuration information.}
\usage{
read_ini(infile)
}
\arguments{
\item{infile}{Character string input file path.}
}
\value{
List of parsed ini-parameters.
}
\description{
Read a Windows-style ini-file that is for configuration information.
}
\note{
Adapted from http://bit.ly/11e4Jh0
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

