% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppa.R
\name{read_ppa_lsm}
\alias{read_ppa_lsm}
\title{Read Zonation post-processing analysis (ppa) result file.}
\usage{
read_ppa_lsm(x)
}
\arguments{
\item{x}{Character string input file path.}
}
\value{
List of 3 data frames:
 1. Most important species in units x
 2. Average proportion remaining over all spp in units
 3. Data fractions in units
}
\description{
Read Zonation post-processing analysis (ppa) result file.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
