% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{read_spp}
\alias{read_spp}
\title{Read Zonation variant specific spp-file.}
\usage{
read_spp(infile)
}
\arguments{
\item{infile}{Character string input file path.}
}
\value{
data.frame of parsed spp data.
}
\description{
Read Zonation variant specific spp-file.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

