\name{require.package}
\alias{require.package}
\title{Requires a given package and if not present installs and loads it.}
\usage{
  require.package(package, ...)
}
\arguments{
  \item{\code{package}}{Character name of a package.}

  \item{\code{...}}{Additional arguments passed on to
  \code{\link{install.packages}}.}
}
\description{
  Requires a given package and if not present installs and
  loads it.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

