% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{require_package}
\alias{require_package}
\title{Requires a given package and if not present installs and loads it.}
\usage{
require_package(package, ...)
}
\arguments{
\item{package}{Character name of a package.}

\item{...}{Additional arguments passed on to \code{\link{install.packages}}.}
}
\description{
Requires a given package and if not present installs and loads it.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
