% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonation.R
\name{run_bat}
\alias{run_bat}
\title{Try to run a given batch (bat) files.}
\usage{
run_bat(bat.file, exe = "zig3")
}
\arguments{
\item{bat.file}{Character string path to a Zonation batch (bat) file.}

\item{exe}{Character string for overriding the default Zonation executable
(default: zig3).}
}
\value{
A logical indicating whether running the batch file was successful.
}
\description{
Try to run a given batch (bat) files.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
