% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{selection_coverage}
\alias{selection_coverage}
\title{Intersection of two coverages.}
\usage{
selection_coverage(x, y, thresholds)
}
\arguments{
\item{x}{Numeric matrix.}

\item{y}{Numeric matrix.}

\item{thresholds}{Numeric vector of thresholds used.}
}
\value{
A list with 2 items:
\item{thresholds}{Correlations between 2 matrices with values above a given threshold.}
\item{total}{Overall correlation between the 2 matrices.}
}
\description{
Calculate how much two coverages (as defined by values greater than a given
threshold in two numeric matrices) overlap.
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{post-processing}
\keyword{ppa}
