\name{set.tutorialdir}
\alias{set.tutorialdir}
\title{Set the package options to point to a correct location of the Zonation
tutorial}
\usage{
  set.tutorialdir(x)
}
\arguments{
  \item{\code{x}}{character string directory path}
}
\description{
  Set the package options to point to a correct location of
  the Zonation tutorial
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

