% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zvariant-methods.R
\docType{methods}
\name{set_dat_param}
\alias{set_dat_param}
\alias{set_dat_param,Zvariant-method}
\title{Set a specified run setting parameter value}
\usage{
set_dat_param(x, parameter, value)

\S4method{set_dat_param}{Zvariant}(x, parameter, value)
}
\arguments{
\item{x}{Zvariant object.}

\item{parameter}{Character string name of the parameter.}

\item{value}{Character string or numeric value of the parameter.}
}
\value{
x Zvariant object.
}
\description{
Essentially these are parameter values corresponding to Zonation dat-files.
This function is used to set values of given parameter. In the current
implementation, sections have no meaning.
}
\note{
Method does not check for the sanity of the values provided.
}
\seealso{
\code{\link[zonator:Zvariant-class]{Zvariant-class}} and
\code{\link{get_dat_param}}.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
