% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{set_tutorialdir}
\alias{set_tutorialdir}
\title{Set the package options to point to a correct location of the Zonation
tutorial}
\usage{
set_tutorialdir(x)
}
\arguments{
\item{x}{Character string directory path.}
}
\description{
Set the package options to point to a correct location of the Zonation
tutorial
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

