% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zvariant-methods.R
\docType{methods}
\name{show,Zvariant-method}
\alias{show,Zvariant-method}
\title{Print Zvariant information.}
\usage{
\S4method{show}{Zvariant}(object)
}
\arguments{
\item{object}{\code{ZVariant} object.}
}
\description{
Generic printing function
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
