% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zproject-methods.R
\name{variants}
\alias{variants}
\alias{variants,Zproject-method}
\title{Get all variants in a Zonation project}
\usage{
variants(x)

\S4method{variants}{Zproject}(x)
}
\arguments{
\item{x}{Zproject object.}
}
\value{
A list of Zvariant objects
}
\description{
Get all variants in a Zonation project
}
\seealso{
\code{\link[zonator:Zproject-class]{Zproject-class}}
  and \code{\link[zonator:Zvariant-class]{Zvariant-class}}
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
