% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{weights}
\alias{weights}
\alias{weights,Zvariant-method}
\title{Get weights of a Zonation variant}
\usage{
weights(x)

\S4method{weights}{Zvariant}(x)
}
\arguments{
\item{x}{Zvariatn object.}
}
\value{
A numeric vector of weights
}
\description{
Get weights of a Zonation variant
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
\code{\link[zonator:Zvariant-class]{Zvariant-class}}
}

