\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,Zcurves,missing-method}
\title{Plot Zonation performance curves.}
\arguments{
  \item{x}{Zcurves object.}

  \item{...}{Additional arguments passed on to the speficic
  plotting functions.}
}
\value{
  ggplot object
}
\description{
  Method itself is only a thin wrapper to functions
  \code{\link{plot_curves}} and
  \code{\link{plot_grp_curves}}.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
  \code{\link{Zvariant-class}}
}

