% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zparameters}
\alias{zparameters}
\title{Get all Zonation run configuration parameters.}
\usage{
zparameters(just_names = FALSE)
}
\arguments{
\item{just_names}{Logical indicating if only the parameter names should be
returned.}
}
\value{
Characted vector of paramter names or a list of (parameter = section)
  structure.
}
\description{
This set of parameters is all that is accepted by Zonation.
}
\note{
Parameters are hard-coded to this package and know nothing
  of potential future developments with Zonation.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
