\docType{methods}
\name{getvariant}
\alias{getvariant}
\alias{getvariant,Zproject,ANY-method}
\alias{names}
\alias{names,Zproject-method}
\alias{nvariants}
\alias{nvariants,Zproject-method}
\alias{opendir}
\alias{opendir,Zproject-method}
\title{getvariant
Get a specified variant in a Zonation project}
\arguments{
  \item{\code{x}}{Zproject object}

  \item{\code{index}}{int or string index defining the
  variant required}

  \item{\code{x}}{Zproject object}

  \item{\code{x}}{object}
}
\value{
  Zvariant object

  int number of variants

  invisible
}
\description{
  getvariant Get a specified variant in a Zonation project

  nvariants Get the number of variants included in a
  Zonation project

  Currently support Windows Explorer (Windows) amd Dolphin
  (Linux/KDE).
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}

  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}

  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
  \code{\link[zonator:AllClasses]{Zproject-class}} and
  \code{\link[zonator:AllClasses]{Zvariant-class}}

  \code{\link{Zproject-class}} and
  \code{\link{Zvariant-class}}

  \code{\link{Zproject-class}} and
  \code{\link{Zvariant-class}}
}

