\docType{methods}
\name{featurenames}
\alias{featurenames}
\alias{featurenames,Zresults-method}
\alias{plot_curves}
\alias{plot_curves,Zresults,missing-method}
\title{Plot Zonation performance curves.}
\arguments{
  \item{x}{Zresults object.}

  \item{...}{Additional arguments passed on to the speficic
  plotting functions.}
}
\value{
  ggplot2 object
}
\description{
  Plot Zonation performance curves.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
  \code{\link{Zvariant-class}}
}

