\name{clean.str}
\alias{clean.str}
\title{clean.str
Clean leading and trailing whitespaces from a given string. Additionally,
all occurrences of multiple whitespaces are replaced with a single
whitespace.}
\usage{
  clean.str(x)
}
\arguments{
  \item{x}{character string}
}
\value{
  A cleaned character string
}
\description{
  clean.str Clean leading and trailing whitespaces from a
  given string. Additionally, all occurrences of multiple
  whitespaces are replaced with a single whitespace.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

