% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zonation.R
\name{create_spp}
\alias{create_spp}
\title{Generate spp_file based on a directory of input rasters.}
\usage{
create_spp(filename = "filelist.spp", weight = 1, alpha = 1, bqp = 1,
  bqp_p = 1, cellrem = 0.25, spp_file_dir,
  spp_file_pattern = "[.+\\\\.[(tif)|(img)]", override_path = NULL)
}
\description{
Generate spp_file based on a directory of input rasters.
}

