% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_tutorialdir}
\alias{get_tutorialdir}
\title{Get the directory of Zonation tutorial.}
\usage{
get_tutorialdir()
}
\value{
path Character path to Zonation tutorial directory.
}
\description{
Get the directory of Zonation tutorial.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
