% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zproject-methods.R
\name{names,Zproject-method}
\alias{names,Zproject-method}
\title{Names of variants in Zproject}
\usage{
\S4method{names}{Zproject}(x)
}
\arguments{
\item{x}{\code{Zproject} object.}
}
\description{
Get the names of all the variants within a given \code{\link{Zproject}}.
}
