% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{nfeatures}
\alias{nfeatures}
\alias{nfeatures,Zproject-method}
\alias{nfeatures,Zvariant-method}
\title{Get the number of feature included in a Zonation variant}
\usage{
nfeatures(x)

\S4method{nfeatures}{Zvariant}(x)
}
\arguments{
\item{x}{Zvariant object.}
}
\value{
int number of variants
}
\description{
Get the number of feature included in a Zonation variant
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
\code{\link{Zvariant-class}}
}

