% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zvariant-methods.R
\name{nfeatures}
\alias{nfeatures}
\alias{nfeatures,Zvariant-method}
\title{Get the number of feature included in a Zonation variant}
\usage{
nfeatures(x)

\S4method{nfeatures}{Zvariant}(x)
}
\arguments{
\item{x}{Zvariant object.}
}
\value{
int number of variants
}
\description{
Get the number of feature included in a Zonation variant
}
\seealso{
\code{\link{Zvariant-class}}
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
