% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zproject-methods.R
\name{nvariants}
\alias{nvariants}
\alias{nvariants,Zproject-method}
\title{Get the number of variants included in a Zonation project}
\usage{
nvariants(x)

\S4method{nvariants}{Zproject}(x)
}
\arguments{
\item{x}{Zproject object.}
}
\value{
int number of variants
}
\description{
Get the number of variants included in a Zonation project
}
\seealso{
\code{\link{Zproject-class}} and \code{\link{Zvariant-class}}
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
