% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zproject-methods.R
\name{opendir}
\alias{opendir}
\alias{opendir,Zproject-method}
\title{Open the directory of a Zproject using the system file browser.}
\usage{
opendir(x)

\S4method{opendir}{Zproject}(x)
}
\arguments{
\item{x}{a \code{\link{Zproject}} object.}
}
\value{
invisible
}
\description{
Currently support Windows Explorer (Windows) amd Dolphin (Linux/KDE).
}
\seealso{
\code{\link{Zproject-class}} and \code{\link{Zvariant-class}}
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
