% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_curves}
\alias{plot_curves}
\title{Create ggplot2 object for performance curves.}
\usage{
plot_curves(
  dat,
  monochrome = FALSE,
  invert.x = FALSE,
  fix.y = FALSE,
  main = "Performance curves",
  legend.title = "Performance",
  groups = FALSE
)
}
\arguments{
\item{monochrome}{logical defining if a monochrome color scheme is used
(default: FALSE).}

\item{invert.x}{logical indicating if the X-axis is printed from
1 ("feature remaining", \code{FALSE}) or 0
("landscape under protection", \code{TRUE}).}

\item{main}{character string title for the plot (deafault: 'Performance
curves').}

\item{legend.title}{character string title for legend (default: 'Performance)
.}

\item{groups}{logical indicating whether the plotted data is groups or not
(default: FALSE).}

\item{x}{data frame of curves of groups curves data.}

\item{fix.x}{logical indicating if y-axis should be fixed to [0, 1] (TRUE) or
not (FALSE).}
}
\value{
ggplot2 object
}
\description{
Method ...
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\keyword{internal}
