\name{read.ppa.lsm}
\alias{read.ppa.lsm}
\title{Read Zonation post-processing analysis (ppa) result file.}
\usage{
  read.ppa.lsm(x)
}
\arguments{
  \item{x}{input file path}
}
\value{
  list of 3 data frames: 1. Most important species in units
  x 2. Average proportion remaining over all spp in units
  3. Data fractions in units
}
\description{
  Read Zonation post-processing analysis (ppa) result file.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

