\docType{methods}
\name{read.ppa.lsm}
\alias{read.ppa.lsm}
\title{read.ppa.lsm
Read Zonation post-processing analysis (ppa) result file and return a list of
3 items:
1. Most important species in units
2. Average proportion remaining over all spp in units
3. Data fractions in units}
\usage{
  read.ppa.lsm()
}
\arguments{
  \item{x}{Zproject object}

  \item{index}{int or string index defining the variant
  required}
}
\value{
  Zvariant object
}
\description{
  read.ppa.lsm Read Zonation post-processing analysis (ppa)
  result file and return a list of 3 items: 1. Most
  important species in units 2. Average proportion
  remaining over all spp in units 3. Data fractions in
  units
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

