% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zvariant-methods.R
\docType{methods}
\name{sppweights}
\alias{sppweights}
\alias{sppweights,Zvariant-method}
\title{Get biodiversity feature weights of a Zonation variant}
\usage{
sppweights(x)

\S4method{sppweights}{Zvariant}(x)
}
\arguments{
\item{x}{Zvariatn object.}
}
\value{
A numeric vector of weights
}
\description{
Get biodiversity feature weights of a Zonation variant
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
\code{\link[zonator:Zvariant-class]{Zvariant-class}}
}

