% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unique_grp_names}
\alias{unique_grp_names}
\title{Get unique group names.}
\usage{
unique_grp_names(x)
}
\arguments{
\item{x}{data frame groups data.}
}
\value{
character vector of unique group names
}
\description{
Method extracts group names directly from group curves data frame header
based on a hard-coded set of prefixes
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\keyword{internal}
