\docType{methods}
\name{get_variant}
\alias{get_variant}
\alias{get_variant,Zproject-method}
\alias{names}
\alias{names,Zproject-method}
\alias{nvariants}
\alias{nvariants,Zproject-method}
\alias{opendir}
\alias{opendir,Zproject-method}
\title{Get a specified variant in a Zonation project}
\arguments{
  \item{x}{Zproject object.}

  \item{index}{int or string index defining the variant
  required.}

  \item{x}{Zproject object.}

  \item{x}{object.}
}
\value{
  Zvariant object

  int number of variants

  invisible
}
\description{
  Get a specified variant in a Zonation project

  Get the number of variants included in a Zonation project

  Currently support Windows Explorer (Windows) amd Dolphin
  (Linux/KDE).
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}

  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}

  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
  \code{\link[zonator:Zproject-class]{Zproject-class}} and
  \code{\link[zonator:Zvariant-class]{Zvariant-class}}

  \code{\link{Zproject-class}} and
  \code{\link{Zvariant-class}}

  \code{\link{Zproject-class}} and
  \code{\link{Zvariant-class}}
}

