\docType{methods}
\name{featurenames}
\alias{featurenames}
\alias{featurenames,Zvariant-method}
\title{Get spp feature names for a class \code{Zvariant} instance.}
\arguments{
  \item{x}{Zvariant object.}
}
\value{
  Character vector of spp feature names.
}
\description{
  Get spp feature names for a class \code{Zvariant}
  instance.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
  \code{\link{Zvariant-class}} \code{\link{groupnames}}
  \code{\link{groups}}
}

